+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix kotlinx.serialization when obfuscated by excluding serializers from proguard.
+
 ## [1.0-alpha15] - 2022-10-13
 
 ### Fixed