+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+- Upgraded dependency on Coil from 1.4.0 to 2.2.2.
+
 ### Fixed
 
 - Fix kotlinx.serialization when obfuscated by excluding serializers from proguard.