+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.0-alpha16] - 2022-10-18
+
 ### Changed
 
 - Reporting: Add a module to report anonymous events to nabla servers to help debug some features like video calls.